+++ wecart/src/main/java/org/zerock/wecart/controller/MypageCartController.java
@@ -1,19 +1,35 @@
 package org.zerock.wecart.controller;
 
+import java.util.HashMap;
+
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.SessionAttributes;
+import org.zerock.wecart.domain.UserVO;
+import org.zerock.wecart.exception.ServiceException;
+import org.zerock.wecart.service.pricecompare.MypageCartService;
 
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import lombok.extern.log4j.Log4j2;
 
 @NoArgsConstructor
 @Log4j2
 
+@SessionAttributes("__AUTH__")
 @RequestMapping("/mypage/cart")
 @Controller
 public class MypageCartController {
+	
+	@Setter(onMethod_=@Autowired)
+	MypageCartService service;
+	
 	@GetMapping("/get")
 	public void get() {
 		log.trace("get() invoked.");
@@ -59,4 +75,40 @@ public class MypageCartController {
 		
 		return "redirect:wishedPrds";
 	} // wishedProdsRemoved   jhwan
+	
+	@PostMapping("/saveGoodsIntoTodayCart")
+	@ResponseBody
+	public String saveGoodsIntoTodayCart(Model model, @RequestBody HashMap<String, Integer> map) throws ServiceException{
+		log.trace("register({}) invoked.", model);
+		log.trace("HashMap<String, Object> param: {}", map);
+		
+//	    String goods_id;
+//		log.trace("goods_id: {}", goods_id);
+		
+		// member_id와 goods_id를 받고
+		UserVO userVO = (UserVO)model.getAttribute("__AUTH__");
+		Integer goods_id = map.get("goods_id");
+		Integer member_id = Integer.parseInt(userVO.getMember_id()); 
+		
+		// 해당 회원이 가지고 있는 TodayCartId를 리턴 
+		Integer todayCartId = this.service.getTodayCartIdOfMember(member_id);
+		
+		// 만약 TodayCartId가 반환되면
+		if(todayCartId != null) {
+			// (1) member_id, goods_id, TodayCart_id로 member_goods_cart 테이블을 생성
+			this.service.saveGoodsIntoTodayCart(member_id, goods_id, todayCartId);
+			// return true;
+		} 
+		// 만약 가지고 있는 TodayCart가 없다면.
+		else {
+			// (1) TodayCart 생성
+			
+			
+			// (2) TodayCart_id를 받고
+			// (3) member_id, goods_id, TodayCart_id로 member_goods_cart 테이블을 생성
+			// return true;
+		} // if - else
+		
+		return "Success";
+	}
 } //end class
