+++ wecart/src/main/java/org/zerock/wecart/service/pricecompare/PriceCompareService.java
@@ -22,13 +22,5 @@ public interface PriceCompareService {
   //4. id에 해당하는 상품 조회
 	public abstract GoodsVO select(Integer goods_id) throws ServiceException;
 	
-	// 멤버 아이디로 저장된 장바구니_id들을 반환
-	public abstract List<Integer> getInstalledCartIdsOfMember(Integer member_id) throws ServiceException;
-	
-	// 멤버 아이디로 TodayCart_id들을 반환
-	public abstract Integer getTodayCartIdOfMember(Integer member_id) throws ServiceException;
-	
-	// member_id, goods_id, cart_id를 받아 물건을 저장
-	public abstract void saveGoodsIntoTodayCart(Integer member_id, Integer goods_id, Integer cart_id) throws ServiceException;
 
 } //end interface
