+++ wecart/src/main/java/org/zerock/wecart/service/pricecompare/PriceCompareServiceImpl.java
@@ -20,13 +20,11 @@ import lombok.extern.log4j.Log4j2;
 @Service
 public class PriceCompareServiceImpl implements PriceCompareService {
 
-	private PriceCompareMapper priceCompareMapper;
-	private MemberGoodsCartMapper memberGoodsCartMapper; 
+	private PriceCompareMapper mapper;
 	
 	@Autowired
-	public PriceCompareServiceImpl(PriceCompareMapper priceCompareMapper, MemberGoodsCartMapper memberGoodsCartMapper) {
-		this.priceCompareMapper = priceCompareMapper;
-		this.memberGoodsCartMapper = memberGoodsCartMapper;
+	public PriceCompareServiceImpl(PriceCompareMapper priceCompareMapper) {
+		this.mapper = priceCompareMapper;
 	}	//Constructor
 
 	@Transactional
@@ -35,7 +33,7 @@ public class PriceCompareServiceImpl implements PriceCompareService {
 		log.trace("getList({}) invoked.", cri);
 		
 		try {
-			return this.priceCompareMapper.selectAll(cri);
+			return this.mapper.selectAll(cri);
 		} catch(Exception e) {
 			throw new ServiceException(e);
 		} //try-catch
@@ -46,7 +44,7 @@ public class PriceCompareServiceImpl implements PriceCompareService {
 		log.trace("select({}) invoked. ", goods_id);
 		
 		try {
-			return this.priceCompareMapper.select(goods_id);
+			return this.mapper.select(goods_id);
 		} catch(Exception e) {
 			
 			throw new ServiceException(e);
@@ -60,7 +58,7 @@ public class PriceCompareServiceImpl implements PriceCompareService {
 		
 		
 		try {
-			return this.priceCompareMapper.selectTotalCount(keyword);
+			return this.mapper.selectTotalCount(keyword);
 		} catch(Exception e) {
 			throw new ServiceException(e);
 		} //try-catch
@@ -71,47 +69,13 @@ public class PriceCompareServiceImpl implements PriceCompareService {
 		log.trace("getSearchList({}) invoked.", cri);
 		
 		try {
-			return this.priceCompareMapper.selectSearch(cri);
+			return this.mapper.selectSearch(cri);
 		} catch(Exception e) {
 			throw new ServiceException(e);
 		} //try-catch
 	} //getSearchList
 
-	@Override
-	public List<Integer> getInstalledCartIdsOfMember(Integer member_id) throws ServiceException{
-		log.trace("getMemberGoodsCartListForMember() invoked. ");
-		
-		try {
-			return this.memberGoodsCartMapper.selecAllInstalledCartOfMember(member_id);
-		}catch(Exception e) {
-			throw new ServiceException(e);
-		}
-		
-	} //getMemberGoodsCartListForMember
 
-	public Integer getTodayCartIdOfMember(Integer member_id) throws ServiceException{
-		log.trace("getTodayCartIdOfMember({}) invoked ", member_id);
-		
-		try {
-			Integer check = this.memberGoodsCartMapper.selectTodayCartOfMember(member_id);
-			log.trace("check:{}", check);
-			return check;
-		}catch(Exception e) {
-			throw new ServiceException(e);
-		}// try-catch
-	} // getTodayCartIdOfMember
-	
-	
-	public void saveGoodsIntoTodayCart(Integer member_id, Integer goods_id, Integer cart_id) throws ServiceException{
-		log.trace("saveGoodsIntoTodayCart({}, {}, {}) invoked. ", member_id, goods_id, cart_id);
-		
-		try {
-			this.memberGoodsCartMapper.insertRowIntoTodayCart(member_id, goods_id, cart_id);
-			
-		}catch(Exception e) {
-			throw new ServiceException(e);
-		} // try - catch;
-	} // saveGoodsIntoTodayCart
 } //end class
