+++ wecart/src/test/java/org/zerock/wecart/service/MypageCartServiceTests.java
@@ -1,0 +1,5 @@
+package org.zerock.wecart.service;
+
+public class MypageCartServiceTests {
+
+}
